feat: combined pty.go and pipe.go into a unified Process implementation
Updated readme to remove redundant information and focus on developer use.
