+++ example/burst/test_burst.go
@@ -14,7 +14,7 @@ func main() {
 	cancellor := consolestream.NewLocalCancellor(5 * time.Second)
 
 	// Create a process that will generate a large burst
-	process := consolestream.NewPipeProcess("go", []string{"run", "cmd/tester/main.go", "--burst-mb=15"}, consolestream.WithCancellor(cancellor))
+	process := consolestream.NewProcess("go", []string{"run", "cmd/tester/main.go", "--burst-mb=15"}, consolestream.WithCancellor(cancellor))
 
 	// Create context with timeout
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
@@ -33,11 +33,11 @@ func main() {
 
 		partCount++
 		switch event := part.Event.(type) {
-		case *consolestream.PipeOutputData:
+		case *consolestream.OutputData:
 			outputCount++
 			totalBytes += len(event.Data)
-			fmt.Printf("Part %d [%s] %s: %d bytes (total: %d bytes, %.2f MB)\n",
-				outputCount, event.Stream.String(), part.Timestamp.Format("15:04:05.000"), len(event.Data), totalBytes, float64(totalBytes)/(1024*1024))
+			fmt.Printf("Part %d [OUTPUT] %s: %d bytes (total: %d bytes, %.2f MB)\n",
+				outputCount, part.Timestamp.Format("15:04:05.000"), len(event.Data), totalBytes, float64(totalBytes)/(1024*1024))
 		case *consolestream.ProcessStart:
 			fmt.Printf("Event %d: Process started (PID: %d)\n", partCount, event.PID)
 		case *consolestream.ProcessEnd:
